requirements.txt
fastapi==0.101.0
pydantic==2.1.1
PyJWT==2.8.0
SQLAlchemy==2.0.19
uvicorn==0.23.2
alembic==1.11.2
pydantic-settings==2.0.2
bcrypt==4.0.1
PyMySQL==1.1.0